+++ example/lib/pages/memory_player_page.dart
@@ -1,0 +1,62 @@
+import 'dart:io';
+
+import 'package:better_player/better_player.dart';
+import 'package:better_player_example/constants.dart';
+import 'package:better_player_example/utils.dart';
+import 'package:flutter/material.dart';
+
+class MemoryPlayerPage extends StatefulWidget {
+  @override
+  _MemoryPlayerPageState createState() => _MemoryPlayerPageState();
+}
+
+class _MemoryPlayerPageState extends State<MemoryPlayerPage> {
+  BetterPlayerController _betterPlayerController;
+
+  @override
+  void initState() {
+    BetterPlayerConfiguration betterPlayerConfiguration =
+        BetterPlayerConfiguration(
+      aspectRatio: 16 / 9,
+      fit: BoxFit.contain,
+    );
+
+    _betterPlayerController = BetterPlayerController(betterPlayerConfiguration);
+    _setupDataSource();
+    super.initState();
+  }
+
+  void _setupDataSource() async {
+    var filePath = await Utils.getFileUrl(Constants.fileTestVideoUrl);
+    File file = File(filePath);
+
+    List<int> bytes = file.readAsBytesSync().buffer.asUint8List();
+    BetterPlayerDataSource dataSource = BetterPlayerDataSource.memory(bytes);
+    _betterPlayerController.setupDataSource(dataSource);
+  }
+
+  @override
+  Widget build(BuildContext context) {
+    return Scaffold(
+      appBar: AppBar(
+        title: Text("Memory player"),
+      ),
+      body: Column(
+        children: [
+          const SizedBox(height: 8),
+          Padding(
+            padding: const EdgeInsets.symmetric(horizontal: 16),
+            child: Text(
+              "Memory player with plays video from bytes.",
+              style: TextStyle(fontSize: 16),
+            ),
+          ),
+          AspectRatio(
+            aspectRatio: 16 / 9,
+            child: BetterPlayer(controller: _betterPlayerController),
+          ),
+        ],
+      ),
+    );
+  }
+}